Tests/LibWeb/Text/expected/wheel-default-action-after-layout-update.txt

PASS
window.scrollY: 100
